src/requirements.txt
sympy~=1.9
numpy~=1.24.3
matplotlib~=3.5.1
scikit-image~=0.22.0
pillow~=10.3.0
open3d~=0.18.0